Feat[Player]: refactor Avatar component and update Player to use new Avatar path
