+++ src/components/Player/Player.tsx
@@ -1,7 +1,7 @@
 import { memo } from "react";
-import { Avatar } from "./Avatar";
-import { PlayerIcon } from "./PlayerIcon";
+import { Avatar } from "../Avatar";
 import Image from "next/image";
+import { ChevronDownIcon, UserCircleIcon } from "@heroicons/react/24/outline";
 
 interface PlayerProps {
   imgSrc?: string;
@@ -14,9 +14,9 @@ export function Player({ imgSrc, userName, rate }: PlayerProps) {
     <button className="flex cursor-pointer items-center gap-x-2 text-left transition-colors hover:text-(--color-hover)">
       <Avatar>
         {imgSrc ? (
-          <Image src={imgSrc} alt="user image" width={500} height={500} />
+          <Image src={imgSrc} alt="user image" width={32} height={32} />
         ) : (
-          <PlayerIcon />
+          <UserCircleIcon className="size-8 stroke-1" />
         )}
       </Avatar>
 
@@ -24,6 +24,8 @@ export function Player({ imgSrc, userName, rate }: PlayerProps) {
         <span className="text-xs">{userName}</span>
         <span className="text-xs text-gray-500">Rate: {rate}</span>
       </div>
+
+      <ChevronDownIcon className="size-4" />
     </button>
   );
 }
