--- src/components/Player/PlayerIcon.tsx
@@ -1,22 +1,0 @@
-export function PlayerIcon() {
-  return (
-    <div className="relative aspect-square rounded-full border-1">
-      <svg
-        className="absolute top-1.5 left-[50%] translate-x-[-50%]"
-        xmlns="http://www.w3.org/2000/svg"
-        width="28"
-        height="28"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="1"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-        <path d="M8 7a4 4 0 1 0 8 0a4 4 0 0 0 -8 0" />
-        <path d="M6 21v-2a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v2" />
-      </svg>
-    </div>
-  );
-}
